requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
duolingo-api==0.5.3
flake8==3.8.4
Flask==1.1.2
googletrans==4.0.0rc1
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hstspreload==2020.12.22
httpcore==0.12.2
httpx==0.16.1
hyperframe==6.0.0
idna==3.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.15.0
requests==2.25.1
rfc3986==1.4.0
sniffio==1.2.0
urllib3==1.26.2
Werkzeug==1.0.1
